+++ react-components/src/pages/FormsPage/FormPage.tsx
@@ -4,12 +4,12 @@ import Layout from '../../components/Layout'
 import Modal from '../../components/Modal'
 
 import InputSelect from './components/InputSelect'
-import InputDate from './components/InputDate'
-import InputName from './components/Input'
+import RadioGroup from './components/RadioGroup'
 import CardsList from './components/CardsList'
+import InputDate from './components/InputDate'
 import FileInput from './components/FileInput'
 import Checkbox from './components/Checkbox'
-import RadioGroup from './components/RadioGroup'
+import InputName from './components/Input'
 import { useFormPage } from './hooks'
 
 import styles from './FormsPage.module.scss'
