+++ react-components/src/pages/FormsPage/components/InputDate/InputDate.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react'
 import { Control, Controller, FieldErrors } from 'react-hook-form'
 
-import { IErrors } from '../../../types'
+import { IErrors } from '../../hooks'
 
 import styles from './InputDate.module.scss'
