--- react-components/src/pages/types.ts
@@ -1,16 +1,0 @@
-export interface ICardData {
-  name: string
-  date: string
-  category: string
-  image: string
-  checkbox: boolean
-  switcher: string
-}
-
-export interface IErrors {
-  name?: string
-  date?: string
-  category?: string
-  file?: string
-  checkbox?: string
-}
